+++ packages/docs/stories/components/button.stories.ts
@@ -3,16 +3,11 @@
 import '../../../components/src/components/button/button';
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
-import docsTokens from '../../../tokens/src/figma-tokens/_docs.json';
-import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../src/helpers/component.js';
+import { storybookDefaults, storybookHelpers, storybookTemplate, generateStoryDescription } from '../../src/helpers/component.js';
 
 const { args: defaultArgs, argTypes } = storybookDefaults('syn-button');
 const { overrideArgs } = storybookHelpers('syn-button');
 const { generateTemplate } = storybookTemplate('syn-button');
-
-const generateStoryDescription = (attributeName: string) => ({
-  story: (docsTokens?.components?.button as Record<string, any>)?.[attributeName]?.description?.value ?? 'No Description',
-});
 
 const meta: Meta = {
   component: 'button',
@@ -20,7 +15,9 @@ const meta: Meta = {
   argTypes,
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
+      description: {
+        component: generateStoryDescription('button', 'default'),
+      },
     },
   },
   title: 'Components/syn-button',
@@ -32,7 +29,9 @@ type Story = StoryObj;
 export const Default = {
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
+      description: {
+        story: generateStoryDescription('button', 'default'),
+      },
     },
   },
   render: (args: any) => generateTemplate({ args }),
@@ -41,7 +40,9 @@ export const Default = {
 export const Variants: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('variant'),
+      description: {
+        story: generateStoryDescription('button', 'variant'),
+      },
     },
   },
   render: () => html`
@@ -58,7 +59,9 @@ export const Variants: Story = {
 export const Sizes: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('size'),
+      description: {
+        story: generateStoryDescription('button', 'size'),
+      },
     },
   },
   render: () => html`
@@ -90,7 +93,9 @@ export const Focus: Story = {
 export const LinkButtons: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('link'),
+      description: {
+        story: generateStoryDescription('button', 'link'),
+      },
     },
   },
   render: () => html`
@@ -108,7 +113,9 @@ export const LinkButtons: Story = {
 export const SettingACustomWidth: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('width'),
+      description: {
+        story: generateStoryDescription('button', 'width'),
+      },
     },
   },
   render: () => html`
@@ -120,59 +127,61 @@ export const SettingACustomWidth: Story = {
 export const PrefixAndSuffixIcons: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('prefix-suffix'),
+      description: {
+        story: generateStoryDescription('button', 'prefix-suffix'),
+      },
     },
   },
   render: () => html`
   <syn-button size="small">
-    <syn-icon slot="prefix" name="gear"></syn-icon>
+    <syn-icon slot="prefix" name="settings"></syn-icon>
     Settings
   </syn-button>
 
   <syn-button size="small">
-    <syn-icon slot="suffix" name="arrow-counterclockwise"></syn-icon>
+    <syn-icon slot="suffix" name="refresh"></syn-icon>
     Refresh
   </syn-button>
 
   <syn-button size="small">
-    <syn-icon slot="prefix" name="link-45deg"></syn-icon>
-    <syn-icon slot="suffix" name="box-arrow-up-right"></syn-icon>
+    <syn-icon slot="prefix" name="link"></syn-icon>
+    <syn-icon slot="suffix" name="launch"></syn-icon>
     Open
   </syn-button>
 
   <br/><br/>
 
   <syn-button>
-    <syn-icon slot="prefix" name="gear"></syn-icon>
+    <syn-icon slot="prefix" name="settings"></syn-icon>
     Settings
   </syn-button>
 
   <syn-button>
-    <syn-icon slot="suffix" name="arrow-counterclockwise"></syn-icon>
+    <syn-icon slot="suffix" name="refresh"></syn-icon>
     Refresh
   </syn-button>
 
   <syn-button>
-    <syn-icon slot="prefix" name="link-45deg"></syn-icon>
-    <syn-icon slot="suffix" name="box-arrow-up-right"></syn-icon>
+    <syn-icon slot="prefix" name="link"></syn-icon>
+    <syn-icon slot="suffix" name="launch"></syn-icon>
     Open
   </syn-button>
 
   <br/><br/>
 
   <syn-button size="large">
-    <syn-icon slot="prefix" name="gear"></syn-icon>
+    <syn-icon slot="prefix" name="settings"></syn-icon>
     Settings
   </syn-button>
 
   <syn-button size="large">
-    <syn-icon slot="suffix" name="arrow-counterclockwise"></syn-icon>
+    <syn-icon slot="suffix" name="refresh"></syn-icon>
     Refresh
   </syn-button>
 
   <syn-button size="large">
-    <syn-icon slot="prefix" name="link-45deg"></syn-icon>
-    <syn-icon slot="suffix" name="box-arrow-up-right"></syn-icon>
+    <syn-icon slot="prefix" name="link"></syn-icon>
+    <syn-icon slot="suffix" name="launch"></syn-icon>
     Open
   </syn-button>
   <style>
@@ -185,7 +194,9 @@ export const PrefixAndSuffixIcons: Story = {
 export const Caret: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('caret'),
+      description: {
+        story: generateStoryDescription('button', 'caret'),
+      },
     },
   },
   render: () => html`
@@ -202,7 +213,9 @@ export const Caret: Story = {
 export const Loading: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('loading'),
+      description: {
+        story: generateStoryDescription('button', 'loading'),
+      },
     },
   },
   render: () => html`
@@ -219,7 +232,9 @@ export const Loading: Story = {
 export const Disabled: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('disabled'),
+      description: {
+        story: generateStoryDescription('button', 'disabled'),
+      },
     },
   },
   render: () => html`