+++ packages/docs/stories/components/textarea.stories.ts
@@ -5,23 +5,20 @@ import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 import { userEvent } from '@storybook/testing-library';
 import { waitUntil } from '@open-wc/testing-helpers';
-import docsTokens from '../../../tokens/src/figma-tokens/_docs.json';
-import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../src/helpers/component.js';
+import { generateStoryDescription, storybookDefaults, storybookHelpers, storybookTemplate } from '../../src/helpers/component.js';
 
 const { args, argTypes } = storybookDefaults('syn-textarea');
 const { overrideArgs } = storybookHelpers('syn-textarea');
 const { generateTemplate } = storybookTemplate('syn-textarea');
-
-const generateStoryDescription = (attributeName: string) => ({
-  story: (docsTokens?.components?.textarea as any)?.[attributeName]?.description?.value ?? 'No Description',
-});
 
 const meta: Meta = {
   args,
   argTypes,
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
+      description: {
+        component: generateStoryDescription('textarea', 'default'),
+      },
     },
   },
   title: 'Components/syn-textarea',
@@ -33,7 +30,9 @@ type Story = StoryObj;
 export const Default = {
   parameters: {
     docs: {
-      description: generateStoryDescription('default'),
+      description: {
+        story: generateStoryDescription('textarea', 'default'),
+      },
     },
   },
   render: (args: any) => generateTemplate({ args }),
@@ -42,7 +41,9 @@ export const Default = {
 export const Labels: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('labels'),
+      description: {
+        story: generateStoryDescription('textarea', 'labels'),
+      },
     },
   },
   render: () => html`<syn-textarea label="Comments"></syn-textarea>`,
@@ -51,7 +52,9 @@ export const Labels: Story = {
 export const HelpText: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('help-text'),
+      description: {
+        story: generateStoryDescription('textarea', 'help-text'),
+      },
     },
   },
   render: () => html`<syn-textarea label="Feedback" help-text="Please tell us what you think."> </syn-textarea>`,
@@ -60,7 +63,9 @@ export const HelpText: Story = {
 export const Rows: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('rows'),
+      description: {
+        story: generateStoryDescription('textarea', 'rows'),
+      },
     },
   },
   render: () => html`
@@ -77,7 +82,9 @@ export const Rows: Story = {
 export const Placeholders: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('placeholder'),
+      description: {
+        story: generateStoryDescription('textarea', 'placeholder'),
+      },
     },
   },
   render: () => html`<syn-textarea placeholder="Type something"></syn-textarea>`,
@@ -86,7 +93,9 @@ export const Placeholders: Story = {
 export const ReadonlyTextareas: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('readonly'),
+      description: {
+        story: generateStoryDescription('textarea', 'readonly'),
+      },
     },
   },
   render: () => html`<syn-textarea value="Read-only content"  readonly></syn-textarea>`,
@@ -119,7 +128,9 @@ export const Focus: Story = {
 export const Disabled: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('disabled'),
+      description: {
+        story: generateStoryDescription('textarea', 'disabled'),
+      },
     },
   },
   render: () => html`<syn-textarea placeholder="Textarea" help-text="Please tell us what you think." label="Label" disabled></syn-textarea>`,
@@ -128,7 +139,9 @@ export const Disabled: Story = {
 export const Sizes: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('size'),
+      description: {
+        story: generateStoryDescription('textarea', 'size'),
+      },
     },
   },
   render: () => html`
@@ -194,7 +207,9 @@ export const Invalid: Story = {
 export const PreventResizing: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('resize'),
+      description: {
+        story: generateStoryDescription('textarea', 'resize'),
+      },
     },
   },
   render: () => html`<syn-textarea resize="none"></syn-textarea>`,
@@ -203,7 +218,9 @@ export const PreventResizing: Story = {
 export const ExpandWithContent: Story = {
   parameters: {
     docs: {
-      description: generateStoryDescription('resize-auto'),
+      description: {
+        story: generateStoryDescription('textarea', 'resize-auto'),
+      },
     },
   },
   render: () => html`<syn-textarea resize="auto" placeholder="Type something"></syn-textarea>`,